LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-origin-traversal-does-not-fire-navigate-expected.txt


FAIL A cross-origin traversal does not fire the navigate event assert_equals: expected 1 but got 0
